Altered script autoloading functions
Switched the order of fileList and dirList to match what is returned by file.Find. Included wildcard "*" so that file.find returns files in the directory specified. Changed the search location from "MOD" to "LUA" to search specifically inside of the LUA folders. Reduced the length of the directory string and concatenated it with the subfolder. Removed an additional string concatenation with "/" for early optimization due to lua's string hashing on string concatenation / creation.